Move KDF() to kdf.go
